Move normpath out of utils
Moves normpath to file_utils, avoding the need
to import whole utils which brings a chain of
own imports.

Generic utils is an kitchensink anti-pattern.
